up : courseLIst가 아닌 item 으로 로직 이동
